+++ jest.config.js
@@ -18,4 +18,6 @@ module.exports = {
 
   // A list of paths to modules that run some code to configure or set up the testing framework before each test.
   setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
+
+  transformIgnorePatterns: ['/node_modules/(?!@storybook)'],
 };