+++ packages/zephyr/CHANGELOG.md
@@ -3,6 +3,183 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.22.3](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.22.2...@air/zephyr@12.22.3) (2021-07-13)
+
+### Bug Fixes
+
+- **@air/zephyr:** add in tx prop to button link ([8b9ba5d](https://github.com/AirLabsTeam/web-core/commit/8b9ba5d229550d30b8a29ace80e1e1fbeeff9b72))
+- **zephyr/buttonlink:** pass onClick to ButtonLink ([41b9769](https://github.com/AirLabsTeam/web-core/commit/41b9769f2e2d8afd3b279a3593a0b62e60485fc7))
+
+## [12.22.2](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.22.1...@air/zephyr@12.22.2) (2021-07-12)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.22.1](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.22.0...@air/zephyr@12.22.1) (2021-07-12)
+
+**Note:** Version bump only for package @air/zephyr
+
+# [12.22.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.21.0...@air/zephyr@12.22.0) (2021-07-10)
+
+### Bug Fixes
+
+- rebasing with main ([0c546e8](https://github.com/AirLabsTeam/web-core/commit/0c546e8206e524c5f0818965f828796555f89787))
+
+### Features
+
+- adds tertiary option to transaction modal ([0fa6aa9](https://github.com/AirLabsTeam/web-core/commit/0fa6aa9bca95b9ad9565bf1b44dcc89c401c9e49))
+- fixing the doc comment ([679359b](https://github.com/AirLabsTeam/web-core/commit/679359b401ad9d8ee9478b16b35d9f88ce5349d0))
+- grammar fixes ([162aac7](https://github.com/AirLabsTeam/web-core/commit/162aac7a0c4d55ba3e6cc875d31fa768bededbf3))
+- remove the tertiary transac modal with button as its redundant ([38a7096](https://github.com/AirLabsTeam/web-core/commit/38a7096d0170ad341b0788f853e751fddd8edb6c))
+
+# [12.21.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.9...@air/zephyr@12.21.0) (2021-07-09)
+
+### Features
+
+- **@air/zephyr:** add in button link ([33b6aff](https://github.com/AirLabsTeam/web-core/commit/33b6aff866799f6a74f094602369a1e1de18e096))
+
+## [12.20.9](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.8...@air/zephyr@12.20.9) (2021-07-09)
+
+### Reverts
+
+- Revert "chore: removed children text wrapper" ([6cb04c1](https://github.com/AirLabsTeam/web-core/commit/6cb04c10a955d734136be68d82a2cc6ec23358f9))
+
+## [12.20.8](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.7...@air/zephyr@12.20.8) (2021-07-07)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.7](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.6...@air/zephyr@12.20.7) (2021-07-06)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.6](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.5...@air/zephyr@12.20.6) (2021-07-06)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.5](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.4...@air/zephyr@12.20.5) (2021-07-06)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.4](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.3...@air/zephyr@12.20.4) (2021-07-02)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.3](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.2...@air/zephyr@12.20.3) (2021-06-29)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.20.2](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.1...@air/zephyr@12.20.2) (2021-06-28)
+
+### Bug Fixes
+
+- **@air/zephyr:** make checkbox active icon optional ([7e2b761](https://github.com/AirLabsTeam/web-core/commit/7e2b7616dd915b5111d68746025d4f3e8422f034))
+
+## [12.20.1](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.20.0...@air/zephyr@12.20.1) (2021-06-28)
+
+### Bug Fixes
+
+- **@air/zephyr:** export checkbox component ([c33a81e](https://github.com/AirLabsTeam/web-core/commit/c33a81e05793393f897d438db36903f035e60bbf))
+
+# [12.20.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.19.0...@air/zephyr@12.20.0) (2021-06-28)
+
+### Bug Fixes
+
+- **@air/zephyr:** remove tx and rename props for checkbox component ([db666cb](https://github.com/AirLabsTeam/web-core/commit/db666cb41ef1996f68cfe14f3f23c22c712d04b1))
+
+### Features
+
+- **@air/zephyr:** add in checkbox component ([7c13852](https://github.com/AirLabsTeam/web-core/commit/7c13852c7fe8362e77b098d065b6059a19179de2))
+
+# [12.19.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.18.0...@air/zephyr@12.19.0) (2021-06-25)
+
+### Features
+
+- fix prop names ([0617020](https://github.com/AirLabsTeam/web-core/commit/06170200251a40c82ab7e917b16df5aa67bc6ca5))
+- for the tooltip - adds an optional tx prop to allow adding custom styles ([8e02444](https://github.com/AirLabsTeam/web-core/commit/8e02444b18343870e943c159c8f8400881c452a9))
+
+# [12.18.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.17.2...@air/zephyr@12.18.0) (2021-06-24)
+
+### Bug Fixes
+
+- **@air/zephyr:** fix icon button to always be a fixed width ([f890b16](https://github.com/AirLabsTeam/web-core/commit/f890b16911c5b52970088289bbf4f21e067ee9ae))
+
+### Features
+
+- update tooltip to have boolean control of arrow ([21c0d68](https://github.com/AirLabsTeam/web-core/commit/21c0d682812fd2da527c254749cd4fe84b69bc61))
+- **@air/zephyr:** update existing 32 text style and add in 36/40/48 sizes ([297f373](https://github.com/AirLabsTeam/web-core/commit/297f373e836fad8cea8c9e159ae1d2bac0aa3681))
+- **@air/zephyr:** update tooltip label type to ReactNode ([9406b5f](https://github.com/AirLabsTeam/web-core/commit/9406b5fe5ab981bddfefbb802f027d26a8ebdcc3))
+
+## [12.17.2](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.17.1...@air/zephyr@12.17.2) (2021-06-23)
+
+### Bug Fixes
+
+- **@air/zephyr:** update editable text box-shadow to be `inset` ([a91bc21](https://github.com/AirLabsTeam/web-core/commit/a91bc21b57cb1c1ff431bb7154799578427edf5c))
+
+## [12.17.1](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.17.0...@air/zephyr@12.17.1) (2021-06-15)
+
+**Note:** Version bump only for package @air/zephyr
+
+# [12.17.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.16.0...@air/zephyr@12.17.0) (2021-06-09)
+
+### Features
+
+- **@air/zephyr:** split up toggle component ([74546b1](https://github.com/AirLabsTeam/web-core/commit/74546b10c6298d712c5237ea423aa17aef2b4abc))
+
+# [12.16.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.15.3...@air/zephyr@12.16.0) (2021-06-08)
+
+### Bug Fixes
+
+- **@air/zephyr:** export toggle component ([78f5692](https://github.com/AirLabsTeam/web-core/commit/78f569240feaec73555183b0b855a78c1456dda5))
+- **@air/zephyr:** expose onClick in menu items ([1297b7b](https://github.com/AirLabsTeam/web-core/commit/1297b7bdbb5a77f095d29807b12f2ac627d1c2ee))
+- **@air/zephyr:** fix ts error in context menu ([c3d4882](https://github.com/AirLabsTeam/web-core/commit/c3d4882237f2834564d578b406f48a9328aa616a))
+
+### Features
+
+- **@air/zephyr:** add in toggle component ([736c20b](https://github.com/AirLabsTeam/web-core/commit/736c20b2463aecea7c81b23b94072c66287cb5bf))
+
+## [12.15.3](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.15.2...@air/zephyr@12.15.3) (2021-06-03)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.15.2](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.15.1...@air/zephyr@12.15.2) (2021-06-03)
+
+**Note:** Version bump only for package @air/zephyr
+
+## [12.15.1](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.15.0...@air/zephyr@12.15.1) (2021-06-03)
+
+**Note:** Version bump only for package @air/zephyr
+
+# [12.15.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.14.2...@air/zephyr@12.15.0) (2021-06-02)
+
+### Bug Fixes
+
+- **@air/zephyr:** update `text-ui-14` line-height to be 20 instead of 21 ([4b2cb95](https://github.com/AirLabsTeam/web-core/commit/4b2cb954a3761888ec2bc4deba843d936f2cafd8))
+
+### Features
+
+- **@air/zephyr:** add in save on blur, and mutliline support ([3159415](https://github.com/AirLabsTeam/web-core/commit/31594150dd8b3f33cd1417f5ece3de25fcb961aa))
+
+## [12.14.2](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.14.1...@air/zephyr@12.14.2) (2021-06-02)
+
+### Bug Fixes
+
+- **@air/zephyr:** updated typing for editable text and expose validationSchema and minLength prop ([fdc12e2](https://github.com/AirLabsTeam/web-core/commit/fdc12e205b5cdeeae31e0ee58f045c3e2967f047))
+
+## [12.14.1](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.14.0...@air/zephyr@12.14.1) (2021-06-02)
+
+### Bug Fixes
+
+- **@air/zephyr:** fix editable text onEditStateChange type to be optional ([4377070](https://github.com/AirLabsTeam/web-core/commit/437707096f17baea8ebb04d32bdc0f29cee029b3))
+
+# [12.14.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.13.0...@air/zephyr@12.14.0) (2021-06-01)
+
+### Bug Fixes
+
+- **@air/zephyr:** fix editable text types ([a542c54](https://github.com/AirLabsTeam/web-core/commit/a542c541f84044053b6c0aa7ddd22f8aae10b256))
+
+### Features
+
+- **@air/zephyr:** add in format callback to editable text component ([2dc311a](https://github.com/AirLabsTeam/web-core/commit/2dc311a6b5c0af27c91c3518de1f74f2e4937291))
+
 # [12.13.0](https://github.com/AirLabsTeam/web-core/compare/@air/zephyr@12.12.1...@air/zephyr@12.13.0) (2021-05-28)
 
 ### Bug Fixes